+++ tests/test_evaluator.py
@@ -155,7 +155,7 @@ class PythonEvaluatorTests(unittest.TestCase):
 
     def test_add_variable_in_context(self):
         self.evaluator._execute_code('a = 1\nassert a == 1', context=self.evaluator.context)
-        self.assertTrue(self.evaluator._evaluate_code('a == 1', context=self.evaluator.context))
+        self.assertTrue(self.evaluator._evaluate_code('a == 1', context={'a': 1}))
 
     @unittest.skip('http://stackoverflow.com/questions/32894942/listcomp-unable-to-access-locals-defined-in-code-called-by-exec-if-nested-in-fun')
     def test_access_outer_scope(self):