+++ CHANGELOG.rst
@@ -8,7 +8,7 @@ Changelog
   implementation.
 - (Added) ``sismic-behave`` CLI now accepts a ``--show-steps`` parameter, which list the steps (equivalent to
   Behave's overriden ``--steps`` parameter).
-- (Added) ``sismic-behave`` now returns an appropiate exit code.
+- (Added) ``sismic-behave`` now returns an appropriate exit code.
 - (Changed) Reorganisation of ``docs/examples``.
 - (Fixed) Coverage data for ``sismic-behave`` takes the initialization step into account (regression
   introduced in 0.21.0).